Move project-index tests from shared tests into plugin suite
