+++ src/plugin/tests/project-index-identifiers.test.js
@@ -4,7 +4,7 @@ import os from "node:os";
 import path from "node:path";
 import test from "node:test";
 
-import { buildProjectIndex } from "../../plugin/src/project-index/index.js";
+import { buildProjectIndex } from "../src/project-index/index.js";
 
 async function writeFile(rootDir, relativePath, contents) {
     const absolutePath = path.join(rootDir, relativePath);
